refactor: Request, Response DTO 리팩토링 및 각 DTO의 일급 컬렉션 DTO 추가
