+++ src/main/java/roomescape/core/dto/reservation/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.core.dto;
+package roomescape.core.dto.reservation;
 
 public class ReservationRequest {
     private String date;
