+++ src/main/java/roomescape/core/dto/reservationtime/ReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.core.dto;
+package roomescape.core.dto.reservationtime;
 
 public class ReservationTimeRequest {
     private String startAt;
