+++ src/main/java/roomescape/core/dto/theme/ThemeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.core.dto;
+package roomescape.core.dto.theme;
 
 public class ThemeRequest {
     private String name;
