+++ src/main/java/roomescape/web/controller/ReservationController.java
@@ -1,7 +1,6 @@
 package roomescape.web.controller;
 
 import java.net.URI;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,8 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.core.dto.ReservationRequest;
-import roomescape.core.dto.ReservationResponse;
+import roomescape.core.dto.reservation.ReservationRequest;
+import roomescape.core.dto.reservation.ReservationResponse;
+import roomescape.core.dto.reservation.ReservationsResponse;
 import roomescape.core.service.ReservationService;
 
 @RestController
@@ -32,7 +32,7 @@ public class ReservationController {
     }
 
     @GetMapping
-    public ResponseEntity<List<ReservationResponse>> findAll() {
+    public ResponseEntity<ReservationsResponse> findAll() {
         return ResponseEntity.ok(reservationService.findAll());
     }
