+++ src/main/java/roomescape/web/controller/ThemeController.java
@@ -1,7 +1,6 @@
 package roomescape.web.controller;
 
 import java.net.URI;
-import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,8 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.core.dto.ThemeRequest;
-import roomescape.core.dto.ThemeResponse;
+import roomescape.core.dto.theme.ThemeRequest;
+import roomescape.core.dto.theme.ThemeResponse;
+import roomescape.core.dto.theme.ThemesResponse;
 import roomescape.core.service.ThemeService;
 
 @RestController
@@ -49,12 +49,12 @@ public class ThemeController {
     }
 
     @GetMapping
-    public ResponseEntity<List<ThemeResponse>> findAll() {
+    public ResponseEntity<ThemesResponse> findAll() {
         return ResponseEntity.ok(themeService.findAll());
     }
 
     @GetMapping("/popular")
-    public ResponseEntity<List<ThemeResponse>> findPopular() {
+    public ResponseEntity<ThemesResponse> findPopular() {
         return ResponseEntity.ok(themeService.findPopularTheme());
     }
